Replace array_intersect_assoc with array_intersect
In my case, I use the default category taxonomy but I had to articles where the 2 categories where in different order. As _assoc uses the array key to compare, it did not match them, so array_intersect seems more appropriate to bypass the order of the taxonomy and match them in any order.